+++ client/src/api/eventstream.tsx
@@ -6,14 +6,13 @@ import { endCombat, newCombat, updateCombat } from "./combatSlice";
 import { updateControlPoint } from "./controlPointsSlice";
 import {
   deselectFlight,
-  registerFlight,
+  registerFlights,
   selectFlight,
-  unregisterFlight,
-  updateFlight,
+  unregisterFlights,
+  updateFlights,
   updateFlightPositions,
 } from "./flightsSlice";
 import {
-  addFrontLine,
   deleteFrontLine,
   updateFrontLine,
 } from "./frontLinesSlice";
@@ -23,14 +22,16 @@ import {
   ControlPoint,
   Flight,
   FrontLine,
+  IadsConnection,
   Tgo,
+  UnculledZone,
 } from "./liberationApi";
 import { navMeshUpdated } from "./navMeshSlice";
 import { updateTgo } from "./tgosSlice";
 import { threatZonesUpdated } from "./threatZonesSlice";
+import { unculledZonesUpdated } from "./unculledZonesSlice";
 import { LatLng } from "leaflet";
-import { updateIadsConnection } from "./iadsNetworkSlice";
-import { IadsConnection } from "./_liberationApi";
+import { updateIadsConnection, removeIadsConnection } from "./iadsNetworkSlice";
 
 interface GameUpdateEvents {
   updated_flight_positions: { [id: string]: LatLng };
@@ -38,18 +39,19 @@ interface GameUpdateEvents {
   updated_combats: Combat[];
   ended_combats: string[];
   navmesh_updates: boolean[];
-  unculled_zones_updated: boolean;
+  updated_unculled_zones: UnculledZone[];
   threat_zones_updated: boolean;
   new_flights: Flight[];
-  updated_flights: string[];
+  updated_flights: Flight[];
   deleted_flights: string[];
   selected_flight: string | null;
   deselected_flight: boolean;
-  new_front_lines: FrontLine[];
-  updated_front_lines: string[];
+  updated_front_lines: FrontLine[];
   deleted_front_lines: string[];
   updated_tgos: string[];
-  updated_control_points: number[];
+  updated_control_points: ControlPoint[];
+  updated_iads: IadsConnection[];
+  deleted_iads: string[];
   reset_on_map_center: LatLng | null;
   game_unloaded: boolean;
   new_turn: boolean;
@@ -87,6 +89,10 @@ export const handleStreamedEvents = (
     });
   }
 
+  if (events.updated_unculled_zones.length > 0) {
+    dispatch(unculledZonesUpdated(events.updated_unculled_zones));
+  }
+
   if (events.threat_zones_updated) {
     dispatch(liberationApi.endpoints.getThreatZones.initiate()).then(
       (result) => {
@@ -97,19 +103,16 @@ export const handleStreamedEvents = (
     );
   }
 
-  for (const flight of events.new_flights) {
-    dispatch(registerFlight(flight));
+  if (events.new_flights.length > 0) {
+    dispatch(registerFlights(events.new_flights));
   }
 
-  for (const id of events.updated_flights) {
-    backend.get(`/flights/${id}?with_waypoints=true`).then((response) => {
-      const flight = response.data as Flight;
-      dispatch(updateFlight(flight));
-    });
+  if (events.updated_flights.length > 0) {
+    dispatch(updateFlights(events.updated_flights));
   }
 
-  for (const id of events.deleted_flights) {
-    dispatch(unregisterFlight(id));
+  if (events.deleted_flights.length > 0) {
+    dispatch(unregisterFlights(events.deleted_flights));
   }
 
   if (events.deselected_flight) {
@@ -120,19 +123,12 @@ export const handleStreamedEvents = (
     dispatch(selectFlight(events.selected_flight));
   }
 
-  for (const front of events.new_front_lines) {
-    dispatch(addFrontLine(front));
+  if (events.updated_front_lines.length > 0) {
+    dispatch(updateFrontLine(events.updated_front_lines));
   }
 
-  for (const id of events.updated_front_lines) {
-    backend.get(`/front-lines/${id}`).then((response) => {
-      const front = response.data as FrontLine;
-      dispatch(updateFrontLine(front));
-    });
-  }
-
-  for (const id of events.deleted_front_lines) {
-    dispatch(deleteFrontLine(id));
+  if (events.deleted_front_lines.length > 0) {
+    dispatch(deleteFrontLine(events.deleted_front_lines));
   }
 
   for (const id of events.updated_tgos) {
@@ -140,18 +136,18 @@ export const handleStreamedEvents = (
       const tgo = response.data as Tgo;
       dispatch(updateTgo(tgo));
     });
-    backend.get(`/iads-network/for-tgo/${id}`).then((response) => {
-      for (const connection of response.data) {
-        dispatch(updateIadsConnection(connection as IadsConnection));
-      }
-    });
   }
 
-  for (const id of events.updated_control_points) {
-    backend.get(`/control-points/${id}`).then((response) => {
-      const cp = response.data as ControlPoint;
-      dispatch(updateControlPoint(cp));
-    });
+  if (events.updated_control_points.length > 0) {
+      dispatch(updateControlPoint(events.updated_control_points));
+  }
+
+  if (events.deleted_iads.length > 0) {
+    dispatch(removeIadsConnection(events.deleted_iads));
+  }
+
+  if (events.updated_iads.length > 0) {
+    dispatch(updateIadsConnection(events.updated_iads));
   }
 
   if (events.reset_on_map_center != null) {