+++ game/layout/__init__.py
@@ -1,4 +1,4 @@
-from .layout import TgoLayout, TgoLayoutGroup
+from .layout import TgoLayout, TgoLayoutGroup, TgoLayoutUnitGroup
 from .layoutloader import LayoutLoader
 
 LAYOUTS = LayoutLoader()