+++ game/missiongenerator/tgogenerator.py
@@ -18,11 +18,19 @@ from dcs.action import DoScript, SceneryDestructionZone
 from dcs.condition import MapObjectIsDead
 from dcs.country import Country
 from dcs.point import StaticPoint
-
+from dcs.ships import (
+    CVN_71,
+    CVN_72,
+    CVN_73,
+    CVN_75,
+    Stennis,
+)
 from dcs.statics import Fortification
 from dcs.task import (
     ActivateBeaconCommand,
     ActivateICLSCommand,
+    ActivateLink4Command,
+    ActivateACLSCommand,
     EPLRS,
     FireAtPoint,
     OptAlarmState,
@@ -33,6 +41,7 @@ from dcs.unit import Unit, InvisibleFARP
 from dcs.unitgroup import MovingGroup, ShipGroup, StaticGroup, VehicleGroup
 from dcs.unittype import ShipType, VehicleType
 from dcs.vehicles import vehicle_map
+from game.missiongenerator.missiondata import CarrierInfo, MissionData
 
 from game.radio.radios import RadioFrequency, RadioRegistry
 from game.radio.tacan import TacanBand, TacanChannel, TacanRegistry, TacanUsage
@@ -343,6 +352,7 @@ class GenericCarrierGenerator(GroundObjectGenerator):
         icls_alloc: Iterator[int],
         runways: Dict[str, RunwayData],
         unit_map: UnitMap,
+        mission_data: MissionData,
     ) -> None:
         super().__init__(ground_object, country, game, mission, unit_map)
         self.ground_object = ground_object
@@ -351,6 +361,7 @@ class GenericCarrierGenerator(GroundObjectGenerator):
         self.tacan_registry = tacan_registry
         self.icls_alloc = icls_alloc
         self.runways = runways
+        self.mission_data = mission_data
 
     def generate(self) -> None:
 
@@ -385,9 +396,22 @@ class GenericCarrierGenerator(GroundObjectGenerator):
                 )
                 tacan_callsign = self.tacan_callsign()
                 icls = next(self.icls_alloc)
-                self.activate_beacons(ship_group, tacan, tacan_callsign, icls)
+                link4 = None
+                if carrier_type in [Stennis, CVN_71, CVN_72, CVN_73, CVN_75]:
+                    link4 = self.radio_registry.alloc_uhf()
+                self.activate_beacons(ship_group, tacan, tacan_callsign, icls, link4)
                 self.add_runway_data(
                     brc or Heading.from_degrees(0), atc, tacan, tacan_callsign, icls
+                )
+                self.mission_data.carriers.append(
+                    CarrierInfo(
+                        group_name=ship_group.name,
+                        unit_name=ship_group.units[0].name,
+                        callsign=tacan_callsign,
+                        freq=atc,
+                        tacan=tacan,
+                        blue=self.control_point.captured,
+                    )
                 )
 
     @property
@@ -406,6 +430,8 @@ class GenericCarrierGenerator(GroundObjectGenerator):
             if self.game.theater.is_in_sea(point):
                 group.points[0].speed = carrier_speed.meters_per_second
                 group.add_waypoint(point, carrier_speed.kph)
+                # Rotate the whole ground object to the new course
+                self.ground_object.rotate(brc)
                 return brc
         return None
 
@@ -414,7 +440,11 @@ class GenericCarrierGenerator(GroundObjectGenerator):
 
     @staticmethod
     def activate_beacons(
-        group: ShipGroup, tacan: TacanChannel, callsign: str, icls: int
+        group: ShipGroup,
+        tacan: TacanChannel,
+        callsign: str,
+        icls: int,
+        link4: Optional[RadioFrequency] = None,
     ) -> None:
         group.points[0].tasks.append(
             ActivateBeaconCommand(
@@ -428,6 +458,11 @@ class GenericCarrierGenerator(GroundObjectGenerator):
         group.points[0].tasks.append(
             ActivateICLSCommand(icls, unit_id=group.units[0].id)
         )
+        if link4 is not None:
+            group.points[0].tasks.append(
+                ActivateLink4Command(int(link4.mhz), group.units[0].id)
+            )
+            group.points[0].tasks.append(ActivateACLSCommand(unit_id=group.units[0].id))
 
     def add_runway_data(
         self,
@@ -582,6 +617,7 @@ class TgoGenerator:
         radio_registry: RadioRegistry,
         tacan_registry: TacanRegistry,
         unit_map: UnitMap,
+        mission_data: MissionData,
     ) -> None:
         self.m = mission
         self.game = game
@@ -591,6 +627,7 @@ class TgoGenerator:
         self.icls_alloc = iter(range(1, 21))
         self.runways: Dict[str, RunwayData] = {}
         self.helipads: dict[ControlPoint, StaticGroup] = {}
+        self.mission_data = mission_data
 
     def generate(self) -> None:
         for cp in self.game.theater.controlpoints:
@@ -618,6 +655,7 @@ class TgoGenerator:
                         self.icls_alloc,
                         self.runways,
                         self.unit_map,
+                        self.mission_data,
                     )
                 elif isinstance(ground_object, LhaGroundObject):
                     generator = LhaGenerator(
@@ -631,6 +669,7 @@ class TgoGenerator:
                         self.icls_alloc,
                         self.runways,
                         self.unit_map,
+                        self.mission_data,
                     )
                 elif isinstance(ground_object, MissileSiteGroundObject):
                     generator = MissileSiteGenerator(
@@ -641,3 +680,4 @@ class TgoGenerator:
                         ground_object, country, self.game, self.m, self.unit_map
                     )
                 generator.generate()
+        self.mission_data.runways = list(self.runways.values())