+++ qt_ui/windows/QLiberationWindow.py
@@ -8,6 +8,7 @@ from PySide2.QtGui import QCloseEvent, QIcon
 from PySide2.QtWidgets import (
     QAction,
     QActionGroup,
+    QApplication,
     QDesktopWidget,
     QFileDialog,
     QMainWindow,
@@ -474,5 +475,7 @@ class QLiberationWindow(QMainWindow):
             self._save_window_geometry()
             super().closeEvent(event)
             self.dialog = None
+            for window in QApplication.topLevelWidgets():
+                window.close()
         else:
             event.ignore()